feat: implement multi-database support and comprehensive testing framework
- Add PostgreSQL, Qdrant, and Redis backend implementations
- Implement comprehensive test suite including unit, integration, and performance tests
- Add benchmark tests for database, encryption, file operations, and workflows
- Create structured test fixtures for adapters, characters, configs, and workflows
- Refactor database layer with improved error handling and backend abstraction
- Update documentation with detailed database connection guide
- Update Docker configuration and Python dependencies
- Remove build artifacts from version control

+++ desktop_app/src-tauri/src/commands/privacy.rs
@@ -279,7 +279,7 @@ pub fn auto_cleanup_expired_data(
 
     // 自动清除旧日志
     if settings.auto_clear_logs_days > 0 {
-        if let Ok(result) = cleanup_manager.cleanup_old_data(settings.auto_clear_logs_days) {
+        if let Ok(result) = cleanup_manager.cleanup_old_data(settings.auto_clear_logs_days as i64) {
             total_result.items_deleted += result.items_deleted;
             total_result.space_freed_bytes += result.space_freed_bytes;
             total_result.errors.extend(result.errors);
